Reincorporated Raylib

+++ src/game/obstacle.cpp
@@ -8,5 +8,5 @@ Obstacle obstacle;
 void obstacleCollisionWall()
 {
 	// Collision logic: player vs walls
-	if (obstacle.x < 0 -(obstacle.radius)) obstacle.x = screenWidth - obstacle.radius;
+	if (obstacle.x < 0 -(obstacle.radius)) obstacle.x = static_cast<int>(screenWidth) - obstacle.radius;
 };
